BOL-375 refactor: terms 관련 model(in data module) package 분리
