+++ data/src/main/java/com/yapp/bol/data/datasource/terms/TermsDataSource.kt
@@ -1,7 +1,7 @@
 package com.yapp.bol.data.datasource.terms
 
-import com.yapp.bol.data.model.auth.TermsRequest
-import com.yapp.bol.data.model.auth.TermsResponse
+import com.yapp.bol.data.model.terms.TermsRequest
+import com.yapp.bol.data.model.terms.TermsResponse
 import com.yapp.bol.domain.model.ApiResult
 import kotlinx.coroutines.flow.Flow
